Open cache interface
